Whoof this was a roller coaster lol
So realized that shelling out meant I couldn't use Webmock and that was
my hint that I was testing this wrong. I was also just testing Thor
internals so switched to using the `start` command from Thor directly
in testing. This was better, but then I realized how I was exiting from
the program was causing errors. Since I used `exit 1` it was killing my
spec run, once I wasn't sub-shelling out. So I changed to raising a
Thor::Error which I believe is more correct. This is an error that
prints out without the stack trace and is meant to be user visible.

Now I can finally start mocking the Webmock interactions and get this
test suite working!