feat: Look at 100 recent builds instead of just 30. (#73)
CircleCI limits the number of returned builds to 30 by default, but the
endpoint supports up to 100. When using CircleCI workflows, it's
entirely possible for a single push to trigger more than 30 builds,
which can result in a behavior where sleet will be unable to find some
of the "overflow" builds.

Bumping this to 100 makes it still possible but much less likely.